requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
cffi==1.11.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.1.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyJWT==1.5.3
pyOpenSSL==17.5.0
pytz==2018.3
queuelib==1.4.2
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
w3lib==1.19.0
zope.interface==4.4.3
